+++ InstanceSearch/InstanceSearch.js
@@ -1,8 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import noop from 'lodash/noop';
+import { FormattedMessage } from 'react-intl';
 
 import { Icon } from '@folio/stripes/components';
+import { useCallout } from '@folio/stripes/core';
+
 import {
   PluginFindRecord,
   PluginFindRecordModal,
@@ -15,6 +18,7 @@ import { getFilterConfig } from '../Imports/imports/filterConfig';
 import useInstancesQuery from '../hooks/useInstancesQuery';
 
 import { CONFIG_TYPES } from '../Imports/imports/constants';
+import { parseHttpError } from '../utils';
 
 const query = {
   query: '',
@@ -29,6 +33,8 @@ const InstanceSearch = ({
   onClose,
   ...rest
 }) => {
+  const callout = useCallout();
+
   const [segment, setSegment] = useState('instances');
   const [instances, setInstances] = useState([]);
   const {
@@ -37,11 +43,12 @@ const InstanceSearch = ({
   } = getFilterConfig(segment);
 
   const results = useInstancesQuery(instances);
-  const isLoading = results.some(result => result.isLoading);
+  const isLoading = results?.isLoading;
+  const isError = results?.isError;
 
   useEffect(() => {
-    if (!isLoading && results.length) {
-      const result = isMultiSelect ? results.map(r => r.data) : results?.[0]?.data;
+    if (!isLoading && !isError && results?.data?.instances?.length) {
+      const result = isMultiSelect ? results.data.instances : results.data.instances[0];
       selectInstance(result);
       setInstances([]);
 
@@ -50,6 +57,23 @@ const InstanceSearch = ({
       }
     }
   }, [isLoading, results, isMultiSelect, selectInstance]);
+
+  useEffect(() => {
+    const getError = async () => {
+      const response = await results?.error.response;
+      const httpError = await parseHttpError(response);
+      const message = httpError?.message ? httpError.message : <FormattedMessage id="ui-plugin-find-instance.communicationProblem" />;
+
+      callout.sendCallout({
+        type: 'error',
+        message,
+      });
+    };
+
+    if (isError) {
+      getError().then();
+    }
+  }, [isError]);
 
   return (
     <PluginFindRecord