+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Remove eslint deps that are already listed in eslint-config-stripes. Refs UIPFI-124.
 * Remove the translation dependency from ui-inventory. Refs UIPFI-24.
+* Fetch instances from mod-search instead of mod-inventory-storage. Refs UIPFI-131.
 
 ## [7.0.0](https://github.com/folio-org/ui-plugin-find-instance/tree/v7.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-instance/compare/v6.5.0...v7.0.0)